% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_binary}
\alias{as_binary}
\title{Converts returns into binary numbers.}
\usage{
as_binary(rets)
}
\arguments{
\item{rets}{zoo or numeric :: A sequence of returns (logarithmic or nominal)}
}
\value{
A zoo object or numeric vector containing the binarized returns.
}
\description{
Positive returns are represented as a 1, negative returns are represented as a 0.
}

