% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_frequency}
\alias{as_frequency}
\title{Converts a zoo object from a daily frequency to a lower frequency.}
\usage{
as_frequency(logrets.zoo, frequency)
}
\arguments{
\item{logrets.zoo}{zoo :: A zoo object of returns at daily frequency.}

\item{frequency}{numeric or string :: A value indicating the desired new frequency. Options are an integer value
greater than or equal to 1 or a string set to "Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun", "Week", or "Month".}
}
\value{
A zoo object containing the returns converted to the specified frequency.
}
\description{
This function converts a zoo object containing daily logarithmic returns into a specified frequency
which is lower than daily e.g. weekly returns or five-daily returns.
}

