% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_interpolated}
\alias{as_interpolated}
\title{Interpolates missing days in the sequence with zero returns.}
\usage{
as_interpolated(logrets.zoo)
}
\arguments{
\item{logrets.zoo}{zoo :: A zoo object of returns at daily frequency.}
}
\value{
A zoo object containing the interpolated time series.
}
\description{
This is required to accurately measure the returns on frequencies defined by day-of-the-week, week, or
month frequencies. This ensures that, for example, Monday to Monday returns are only ever one week long.
}

