% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_levels}
\alias{as_levels}
\title{Converts log returns to prices}
\usage{
as_levels(logrets)
}
\arguments{
\item{logrets}{zoo or numeric :: A sequence of log returns for some time series.}
}
\value{
A zoo object or numeric vector containing price levels.
}
\description{
This function converts log returns to a price series.
}

