% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_logreturns}
\alias{as_logreturns}
\title{Converts prices to log returns.}
\usage{
as_logreturns(levels, order = 1)
}
\arguments{
\item{levels}{zoo or numeric :: A sequence of prices for some time series}

\item{order}{numeric :: The order of the returns to compute. When 1 this is the same as daily returns.}
}
\value{
A zoo object or numeric vector containing log returns of order, order.
}
\description{
This function converts prices to log returns of order, order.
}

