% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_residuals_lm}
\alias{as_residuals_lm}
\title{Returns the residual of the returns and a linear model fitted to the returns lagged by one time period.}
\usage{
as_residuals_lm(rets)
}
\arguments{
\item{rets}{zoo or numeric vector :: nominal returns of the asset}
}
\value{
A zoo object or numeric vector containing the residuals from a regression.
}
\description{
This function returns the residual of the returns and a linear model fitted to the returns lagged by
one time period. It is a common procedure used in many statistical tests of randomness.
}

