% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_residuals_rolling}
\alias{as_residuals_rolling}
\title{Converts returns to residual returns above a rolling simple mean.}
\usage{
as_residuals_rolling(rets, w = 126)
}
\arguments{
\item{rets}{zoo or numeric :: nominal returns of the asset}

\item{w}{numeric :: the length of the simple-moving-average to compute.}
}
\value{
A zoo object or numeric vector containing nominal residual returns.
}
\description{
This function extracts a rolling simple moving average from the returns.
}

