% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_returns}
\alias{as_returns}
\title{Converts prices to nominal returns.}
\usage{
as_returns(levels, order = 1)
}
\arguments{
\item{levels}{zoo or numeric :: A sequence of prices for some time series.}

\item{order}{numeric :: The order of the returns to compute.}
}
\value{
A zoo object or numeric vector containing returns of order, order.
}
\description{
This function converts price levels to nominal returns of order, order.
}

