% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{as_trend_rolling}
\alias{as_trend_rolling}
\title{Returns the w-day simple moving average of the returns.}
\usage{
as_trend_rolling(rets, w = 126)
}
\arguments{
\item{rets}{zoo or numeric vector :: nominal returns of the asset}

\item{w}{numeric :: the length of the simple-moving-average to compute}
}
\value{
A zoo object or numeric vector containing nominal rolling residual returns.
}
\description{
This returns the w-day simple moving average return of the returns.
}

