% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestIndependentRuns.R
\name{count_ones}
\alias{count_ones}
\title{Count the number of ones in a binary sequence}
\usage{
count_ones(binrets)
}
\arguments{
\item{binrets}{numeric :: A binary sequence e.g. c(0, 1, 0, 0, 1, 1)}
}
\description{
Returns the count of the number of one symbols in a binary sequence.
}

