% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestIndependentRuns.R
\name{count_zeroes}
\alias{count_zeroes}
\title{Count the number of zeroes in a binary sequence}
\usage{
count_zeroes(binrets)
}
\arguments{
\item{binrets}{numeric :: A binary sequence e.g. c(0, 1, 0, 0, 1, 1)}
}
\description{
Returns the count of the number of zero symbols in a binary sequence.
}

