% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDownloaders.R
\name{data_quandl_downloader}
\alias{data_quandl_downloader}
\title{Download a set of codes from Quandl.com.}
\usage{
data_quandl_downloader(codes)
}
\arguments{
\item{codes}{character :: a list of characters containing codes.}
}
\value{
zoo :: A zoo object containing daily prices for each code.
}
\description{
This function downloads a set of codes form Quandl.com. The best variable for each one of the datasets
is extracted by defaults. For equity time series this usually corresponds to the adjusted close.
}

