% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetInterfaces.R
\name{data_quandl_indices}
\alias{data_quandl_indices}
\title{A list of global stock market indices available on Quandl.com}
\usage{
data_quandl_indices()
}
\description{
This function returns a vector of names corresponding to the codes of various global stock market
indices available for free on Quandl.com.
}

