% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestMarketEfficiency.R
\name{is.random}
\alias{is.random}
\title{Is your time series random?}
\usage{
is.random(S, a = 0.99, freqs1 = c(1, 2, 3, 4, 5), freqs2 = c("Mon", "Tue",
  "Wed", "Thu", "Fri", "Week", "Month"))
}
\arguments{
\item{S}{zoo :: A price series. MUST be a zoo object (for downsampling to work).}

\item{a}{numeric :: Controls the significance level of the tests e.g. 0.95 for 95 percent.}

\item{freqs1}{numeric :: A vector containing the lags to use for the tests.}

\item{freqs2}{character :: A vector containing time-based downsampling frequencies.}
}
\description{
This function performs a number of randomness tests on price data at different frequencies.
}
\details{
TODO
}

