% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestMarketEfficiency.R
\name{is_random}
\alias{is_random}
\title{Test whether or not a price series is random.}
\usage{
is_random(S, a = 0.99, freqs1 = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
  freqs2 = c("Mon", "Tue", "Wed", "Thu", "Fri", "Week", "Month"))
}
\arguments{
\item{S}{zoo :: A price series. MUST be a zoo object (for downsampling to work).}

\item{a}{numeric :: Controls the significance level of the tests e.g. 0.95 for 95 percent.}

\item{freqs1}{numeric :: A vector containing the lags to use for the tests.}

\item{freqs2}{character :: A vector containing time-based downsampling frequencies.}
}
\value{
data.frame :: A data.frame containing all of the tabulated test results.
}
\description{
This function performs a number of randomness tests on price data at different frequencies.
}
\details{
This function accepts a price series, S, and transforms it into a set of price changes over different
frequencies. The first set of frequencies specifies the order over which to compute the price changes whereas
the second set of frequencies specifies a point-in-time interval over which to compute the price changes. For
example if freqs1 = c(1, 2) the every-one-day and every-two-days price changes will be computed and if freqs2 =
c("Wed", "Week") the wednesday-to-wednesday and week-to-week price changes will be computed. Each set of price
changes is then fed through the set of all randomness tests given a confidence interval, a. The results of each
test - including the test statistic, two-sided p-value, sample size, and result - are tabulated and returned.
}

