% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_fattails}
\alias{plot_fattails}
\title{This function plots the distribution of the returns.}
\usage{
plot_fattails(returns)
}
\arguments{
\item{returns}{zoo or numeric vector :: The returns you want to plot the distribution of.}
}
\description{
This function basically just fits a density kernel to the returns and plots it. This will allow you to
see whether or not the returns you are modelling or testing are fat-tailed or not.
}

