% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_results}
\alias{plot_results}
\title{A handy function for plotting the results of the randomness tests by frequency and test name.}
\usage{
plot_results(results_df)
}
\arguments{
\item{results_df}{data.frame :: A data.fram produced by is_random.}
}
\description{
This function generates two plots. The first plot is the breakdown of the percentage of randomness
tests which indicated that the return series was non-random by frequency. The second plot is the breakdown of the
percentage of randomness tests which indicated that the return series was non-random by the type of the test.
}

