% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plot_rollvol}
\alias{plot_rollvol}
\title{Plots the rolling volatility of a daily time series.}
\usage{
plot_rollvol(returns)
}
\arguments{
\item{returns}{zoo or numeric vector :: the return series you want to visualize the stochastic volatility of.}
}
\description{
Financial time series are characterized by fat-tails, and stochastic volatility. This plot will plot
the rolling annualized volatility of a daily return series.
}

