% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBrownianMotion.R
\name{simulate_brownian_motion}
\alias{simulate_brownian_motion}
\title{Brownian Motion with or without Drift.}
\usage{
simulate_brownian_motion(n = 252, dt = 0.003968254, vol = 0.25,
  drift = 0)
}
\arguments{
\item{n}{numeric :: The amount of time to simulate.}

\item{dt}{numeric :: Delta t, change in time.}

\item{vol}{numeric :: Annualized volatility.}

\item{drift}{numeric :: Average daily log return.}
}
\value{
A numeric vector containing simulated log returns.
}
\description{
Simulates a logarithmic return series samples from a Brownian Motion model with or without drift.
}
\details{
This function simulates a logarithmic return series from a Brownian Motion model with or without drift.
Brownian Motion was first proposed as a model of security prices by Louis Bachelier in his 1900 PhD thesis entitled
"The Theory of Speculation". Geometric Brownian Motion, Brownian Motion with drift, is commonly used to price
financial securities. It is the model of security returns assumed by the Black Scholes options pricing model.

\itemize{
 \item{"Theorie de la speculation."}{Bachelier, Louis. Gauthier-Villars, 1900.}
}
}

