% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimJumpDiffusion.R
\name{simulate_merton_model}
\alias{simulate_merton_model}
\title{A Merton Jump Diffusion Model}
\usage{
simulate_merton_model(n = 252, dt = 0.003968254, vol = 0.25, drift = 0,
  jlambda = 0.0156, jsigma = 0.01, jmu = -0.0075)
}
\arguments{
\item{n}{numeric :: The amount of time to simulate.}

\item{dt}{numeric :: Delta t, change in time.}

\item{vol}{numeric :: Annualized volatility.}

\item{drift}{numeric :: Average daily log return.}

\item{jlambda}{numeric :: The probability of a jump happening.}

\item{jsigma}{numeric :: The volatility of jumps which occur.}

\item{jmu}{numeric :: The average jump size which occurs.}
}
\value{
A numeric vector containing simulated log returns.
}
\description{
Simulates a logarithmic returns series from a Merton jump diffusion model with drift.
}
\details{
This function simulates a logarithmic return series from a Merton Jump Diffusion model. The Merton Jump
Diffusion model is a popular stochastic process used for simulating time series with fat tails. Which side of the
tail is fat is controlled by the parameters jlambda, jmu, and jsigma.
}

