% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestDurbinWatson.R
\name{test.durbinwatson}
\alias{test.durbinwatson}
\title{Durbin-Watson test for serial correlation.}
\usage{
test.durbinwatson(rets, a = 0.99, v = FALSE)
}
\arguments{
\item{rets}{:: time series returns. May be a zoo or numeric vector.}

\item{a}{:: alpha. This controls the significance level of the results.}

\item{v}{:: verbose? If true, then some additional information is printed out.}
}
\description{
Performs the Durbin-Watson test for serial correlation.
}
\details{
TODO
}

