% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestIndependentRuns.R
\name{test.runs}
\alias{test.runs}
\title{Runs Test for Randomness}
\usage{
test.runs(rets, a = 0.99, v = FALSE)
}
\arguments{
\item{rets}{:: time series returns. May be a zoo or numeric vector.}

\item{a}{:: alpha. This controls the significance level of the results.}

\item{v}{:: verbose? If true, then some additional information is printed out.}
}
\description{
Performs the runs test on binarized returns.
}
\details{
TODO
}

