% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestBartellRank.R
\name{test_bartellrank}
\alias{test_bartellrank}
\title{Bartell's rank test for serial correlation.}
\usage{
test_bartellrank(rets, a = 0.99)
}
\arguments{
\item{rets}{numeric :: time series returns. May be a zoo or numeric vector.}

\item{a}{numeric :: alpha. This controls the significance level of the results.}
}
\description{
Performs the Breusch-Godfrey test for serial correlation.
}

