% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestBreuschGodfrey.R
\name{test_breuschgodfrey}
\alias{test_breuschgodfrey}
\title{Breusch-Godfrey test for serial correlation.}
\usage{
test_breuschgodfrey(rets, a = 0.99)
}
\arguments{
\item{rets}{numeric :: time series returns. May be a zoo or numeric vector.}

\item{a}{numeric :: alpha. This controls the significance level of the results.}
}
\description{
Performs the Breusch-Godfrey test for serial correlation.
}

