% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestDurbinWatson.R
\name{test_durbinwatson}
\alias{test_durbinwatson}
\title{Durbin-Watson test for serial correlation.}
\usage{
test_durbinwatson(rets, a = 0.99)
}
\arguments{
\item{rets}{numeric :: time series returns. May be a zoo or numeric vector.}

\item{a}{numeric :: alpha. This controls the significance level of the results.}
}
\description{
Performs the Durbin-Watson test for serial correlation.
}

