% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestLjungBox.R
\name{test_ljungbox}
\alias{test_ljungbox}
\title{Ljung-Box portmanteau test that autocorrelations are non-zero.}
\usage{
test_ljungbox(rets, a = 0.99, n.lags = 15)
}
\arguments{
\item{rets}{numeric :: time series returns. May be a zoo or numeric vector.}

\item{a}{numeric :: alpha. This controls the significance level of the results.}

\item{n.lags}{numeric :: the number of lags to test up to.}
}
\description{
Performs a Ljung-Box to see if autocorrelations are non-zero.
}

