% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestIndependentRuns.R
\name{test_runs}
\alias{test_runs}
\title{Runs Test for Randomness}
\usage{
test_runs(rets, a = 0.99)
}
\arguments{
\item{rets}{numeric :: time series returns. May be a zoo or numeric vector.}

\item{a}{numeric :: alpha. This controls the significance level of the results.}
}
\description{
Performs the runs test on binarized returns.
}

