% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestIndependentRuns.R
\name{test_runs_longest}
\alias{test_runs_longest}
\title{Computes the length of the longest independent run in a binary sequence.}
\usage{
test_runs_longest(binrets)
}
\arguments{
\item{binrets}{numeric :: A binary sequence e.g. c(0, 1, 0, 0, 1, 1)}
}
\description{
Returns the length of the longest run in a binary sequence.
}

