% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestIndependentRuns.R
\name{test_runs_number}
\alias{test_runs_number}
\title{Computes the number of independent runs in a binary sequence.}
\usage{
test_runs_number(binrets)
}
\arguments{
\item{binrets}{numeric :: A binary sequence e.g. c(0, 1, 0, 0, 1, 1)}
}
\description{
Returns the number of runs in a binary sequence.
}

