% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestVarianceRatioLoMac.R
\name{test_vratio_lo_mac}
\alias{test_vratio_lo_mac}
\title{Given a log price process, X, compute the Z-score which can be used
to accept or reject the hypothesis that the process evolved according to a
Brownian Motion model with drift and stochastic volatility.}
\usage{
test_vratio_lo_mac(rets, a = 0.99, q = 2)
}
\arguments{
\item{rets}{vector :: A log returns process.}

\item{a}{numeric :: The confidence interval to use.}

\item{q}{int :: The sampling interval for the estimator.}
}
\description{
Given a log price process, X, and a sampling interval, q, this
method returns a Z score indicating the confidence we have that X evolved
according to a Brownian Motion mode with drift and stochastic volatility. This
heteroskedasticity-consistent variance ratio test essentially checks to see
whether or not the observed Mr statistic for the number of observations, is
within or out of the limiting distribution defined by the Asymptotic Variance.
}

